literally just added a blank line
No real changes just wanted to make note that this should be used in an editor such as IDLE only in for the display properly. This was great fun to work on, but was merely a school project. Will probably be the last update unless I have the chance to come back and clean this up. However I may have something new involving character generation in the not too distant future.